4.1_AWS/EC2_backend/sayhellotobike/requirements.txt
Django==3.2.8
djangorestframework==3.14.0
django-cors-headers==3.9.0
djangorestframework_simplejwt==5.2.2
psycopg2-binary==2.9.1
Pillow==6.2.2
olefile
gunicorn==20.1.0